+++ ts-client/constants.ts
@@ -7,7 +7,7 @@
 // ---------------------------------------------------------------------------
 
 export const WS_PROTOCOL_VERSION = 1;
-export const WS_CLIENT_VERSION = '2.4.0';
+export const WS_CLIENT_VERSION = '2.4.1';
 
 // ---------------------------------------------------------------------------
 // Connection Constants